update bQ tests
